pyproject.toml
[tool.poetry]
name = "minecraft-server-discord-bot"
version = "0.1.0"
description = "Start your Minecraft server from a Discord bot."
readme = "README.md"
license = "GPL-3.0-only"
authors = ["nottherectifier <nottherectifier@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
mcrcon = "^0.7.0"
disnake = "^2.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
taskipy = "^1.9.0"
python-dotenv = "^0.19.2"

[tool.taskipy.tasks]
bot = { cmd = "python -m bot", help = "Runs Bot"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
